break down commands.rs into multiple files
