+++ src/main.rs
@@ -17,6 +17,9 @@ use error::{Error, Result};
 use opts::{Opts, Subcommand};
 use utils::{QuestionAction, QuestionPolicy};
 
+// Used in BufReader and BufWriter to perform less syscalls
+const BUFFER_CAPACITY: usize = 1024 * 32;
+
 /// The status code returned from `ouch` on error
 pub const EXIT_FAILURE: i32 = libc::EXIT_FAILURE;
